norm done for builtins + fixed specific case in echo
